Profile Scraper Tool/requirements.txt
http.client
requests
python-dotenv
apify-client
langchain
langchain-community